+++ tests/Earth/Move/EndOfDayTest.php
@@ -31,7 +31,7 @@ class EndOfDayTest extends TestCase
         }));
     }
 
-    public function cases(): array
+    public static function cases(): array
     {
         return [
             ['2016-02-29 13:12:11.675', '2016-02-29 23:59:59.999000'],