+++ tests/Earth/Period/MillisecondTest.php
@@ -14,7 +14,7 @@ class MillisecondTest extends TestCase
 {
     public function testInterface()
     {
-        $period = new Millisecond(20);
+        $period = Millisecond::of(20);
 
         $this->assertInstanceOf(Period::class, $period);
         $this->assertSame(0, $period->years());
@@ -100,7 +100,7 @@ class MillisecondTest extends TestCase
         );
     }
 
-    public function cases()
+    public static function cases()
     {
         return [
             [20, 0, 0, 0, 0, 20],