+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## 3.3.0 - 2023-07-08
+
+### Added
+
+- `Innmind\TimeContinuum\Earth\PointInTime\Now::elapsedSince()` now use the high resolution time api when comparing to another instance of `Now` (to avoid errors when time _goes back_ due to the system clock re-synchronisation)
+- `Innmind\TimeContinuum\Earth\Period\Year::of()`
+- `Innmind\TimeContinuum\Earth\Period\Second::of()`
+- `Innmind\TimeContinuum\Earth\Period\Month::of()`
+- `Innmind\TimeContinuum\Earth\Period\Minute::of()`
+- `Innmind\TimeContinuum\Earth\Period\Millisecond::of()`
+- `Innmind\TimeContinuum\Earth\Period\Hour::of()`
+- `Innmind\TimeContinuum\Earth\Period\Day::of()`
+
+### Changed
+
+- Require `innmind/black-box` `5`
+
+### Removed
+
+- Support for PHP `8.1`
+
 ## 3.2.0 - 2023-02-11
 
 ### Added